requirements.txt
requests>=2.31.0
playwright>=1.41.0
googlesearch-python>=1.2.3
configparser>=6.0.0
beautifulsoup4>=4.12.0
shodan>=1.31.0
python-dateutil>=2.8.2
argparse>=1.4.0 